cheri: wireless: Apply the multicast.cocci script to update the ABI
Apply the multicast.cocci script to update the function definitions
for the prepare_multicast and configure_filter scripts.

In order to catch prototypes in header files and implementations
that are in a different file first search for all _directories_
continaing at least one file that contains the strings prepare_multicast
and/or configure_filter.

Then run coccinelle on each directory <DIR> like this:
	spatch \
		--sp-file scripts/coccinelle/cheri/multicast.cocci \
		--macro-file scripts/coccinelle/cheri/multicast.h \
		<DIR>/*.[ch]

+++ drivers/net/wireless/ath/ath12k/mac.h
@@ -235,7 +235,7 @@ ath12k_mac_op_change_vif_links
 void ath12k_mac_op_configure_filter(struct ieee80211_hw *hw,
 				    unsigned int changed_flags,
 				    unsigned int *total_flags,
-				    u64 multicast);
+				    uintptr_t multicast);
 int ath12k_mac_op_hw_scan(struct ieee80211_hw *hw,
 			  struct ieee80211_vif *vif,
 			  struct ieee80211_scan_request *hw_req);
