+++ drivers/net/wireless/ath/ath5k/mac80211-ops.c
@@ -315,7 +315,7 @@ ath5k_bss_info_changed(struct ieee80211_hw *hw, struct ieee80211_vif *vif,
 }
 
 
-static u64
+static uintptr_t
 ath5k_prepare_multicast(struct ieee80211_hw *hw,
 			struct netdev_hw_addr_list *mc_list)
 {
@@ -365,7 +365,7 @@ ath5k_prepare_multicast(struct ieee80211_hw *hw,
  */
 static void
 ath5k_configure_filter(struct ieee80211_hw *hw, unsigned int changed_flags,
-		       unsigned int *new_flags, u64 multicast)
+		       unsigned int *new_flags, uintptr_t multicast)
 {
 #define SUPPORTED_FIF_FLAGS \
 	(FIF_ALLMULTI | FIF_FCSFAIL | \
