+++ drivers/net/wireless/ath/wcn36xx/main.c
@@ -427,7 +427,7 @@ static int wcn36xx_config(struct ieee80211_hw *hw, int radio_idx, u32 changed)
 
 static void wcn36xx_configure_filter(struct ieee80211_hw *hw,
 				     unsigned int changed,
-				     unsigned int *total, u64 multicast)
+				     unsigned int *total, uintptr_t multicast)
 {
 	struct wcn36xx_hal_rcv_flt_mc_addr_list_type *fp;
 	struct wcn36xx *wcn = hw->priv;
@@ -455,7 +455,7 @@ static void wcn36xx_configure_filter(struct ieee80211_hw *hw,
 	kfree(fp);
 }
 
-static u64 wcn36xx_prepare_multicast(struct ieee80211_hw *hw,
+static uintptr_t wcn36xx_prepare_multicast(struct ieee80211_hw *hw,
 				     struct netdev_hw_addr_list *mc_list)
 {
 	struct wcn36xx_hal_rcv_flt_mc_addr_list_type *fp;
