+++ drivers/net/wireless/broadcom/b43legacy/main.c
@@ -2877,7 +2877,8 @@ static void b43legacy_op_bss_info_changed(struct ieee80211_hw *hw,
 
 static void b43legacy_op_configure_filter(struct ieee80211_hw *hw,
 					  unsigned int changed,
-					  unsigned int *fflags,u64 multicast)
+					  unsigned int *fflags,
+					  uintptr_t multicast)
 {
 	struct b43legacy_wl *wl = hw_to_b43legacy_wl(hw);
 	struct b43legacy_wldev *dev = wl->current_dev;
