+++ drivers/net/wireless/broadcom/brcm80211/brcmsmac/mac80211_if.c
@@ -740,7 +740,7 @@ brcms_ops_bss_info_changed(struct ieee80211_hw *hw,
 static void
 brcms_ops_configure_filter(struct ieee80211_hw *hw,
 			unsigned int changed_flags,
-			unsigned int *total_flags, u64 multicast)
+			unsigned int *total_flags, uintptr_t multicast)
 {
 	struct brcms_info *wl = hw->priv;
 	struct bcma_device *core = wl->wlc->hw->d11core;
