+++ drivers/net/wireless/mediatek/mt76/mt7603/main.c
@@ -243,7 +243,7 @@ mt7603_config(struct ieee80211_hw *hw, int radio_idx, u32 changed)
 
 static void
 mt7603_configure_filter(struct ieee80211_hw *hw, unsigned int changed_flags,
-			unsigned int *total_flags, u64 multicast)
+			unsigned int *total_flags, uintptr_t multicast)
 {
 	struct mt7603_dev *dev = hw->priv;
 	u32 flags = 0;
