+++ drivers/net/wireless/ti/wl1251/main.c
@@ -715,7 +715,7 @@ struct wl1251_filter_params {
 	u8 mc_list[ACX_MC_ADDRESS_GROUP_MAX][ETH_ALEN];
 };
 
-static u64 wl1251_op_prepare_multicast(struct ieee80211_hw *hw,
+static uintptr_t wl1251_op_prepare_multicast(struct ieee80211_hw *hw,
 				       struct netdev_hw_addr_list *mc_list)
 {
 	struct wl1251_filter_params *fp;
@@ -756,7 +756,8 @@ static u64 wl1251_op_prepare_multicast(struct ieee80211_hw *hw,
 
 static void wl1251_op_configure_filter(struct ieee80211_hw *hw,
 				       unsigned int changed,
-				       unsigned int *total, u64 multicast)
+				       unsigned int *total,
+				       uintptr_t multicast)
 {
 	struct wl1251_filter_params *fp = (void *)(unsigned long)multicast;
 	struct wl1251 *wl = hw->priv;
