+++ drivers/net/wireless/ti/wlcore/main.c
@@ -3209,7 +3209,7 @@ struct wl1271_filter_params {
 	u8 mc_list[ACX_MC_ADDRESS_GROUP_MAX][ETH_ALEN];
 };
 
-static u64 wl1271_op_prepare_multicast(struct ieee80211_hw *hw,
+static uintptr_t wl1271_op_prepare_multicast(struct ieee80211_hw *hw,
 				       struct netdev_hw_addr_list *mc_list)
 {
 	struct wl1271_filter_params *fp;
@@ -3245,7 +3245,8 @@ static u64 wl1271_op_prepare_multicast(struct ieee80211_hw *hw,
 
 static void wl1271_op_configure_filter(struct ieee80211_hw *hw,
 				       unsigned int changed,
-				       unsigned int *total, u64 multicast)
+				       unsigned int *total,
+				       uintptr_t multicast)
 {
 	struct wl1271_filter_params *fp = (void *)(unsigned long)multicast;
 	struct wl1271 *wl = hw->priv;
