+++ drivers/net/wireless/virtual/mac80211_hwsim.c
@@ -2510,7 +2510,8 @@ static int mac80211_hwsim_config(struct ieee80211_hw *hw, int radio_idx,
 
 static void mac80211_hwsim_configure_filter(struct ieee80211_hw *hw,
 					    unsigned int changed_flags,
-					    unsigned int *total_flags,u64 multicast)
+					    unsigned int *total_flags,
+					    uintptr_t multicast)
 {
 	struct mac80211_hwsim_data *data = hw->priv;
